Add config connect mongoDB
